+++ src/shared/services/api/pessoas/PessoasService.ts
@@ -5,14 +5,16 @@ export interface IListagemPessoa{
     id: number;
     email: string;
     cidadeId: number;
-    nomeCompleto: string;
+    nome: string;
+    sobrenome?: string;
 }
 
 export interface IDetalhePessoa{
     id: number;
     email: string;
     cidadeId: number;
-    nomeCompleto: string;
+    nome: string;
+    sobrenome?: string;
 }
 
 type TPessoasComTotalCount = {
@@ -22,8 +24,11 @@ type TPessoasComTotalCount = {
 
 const getAll = async (page = 1, filter = ''): Promise<TPessoasComTotalCount | Error> => {
     try {
-        const urlRelativa = `/pessoas?_page=${page}&_limit=${Environment.LIMITE_DE_LINHAS}&nomeCompleto_like=${filter}`;
-        const { data, headers } = await Api.get(urlRelativa);
+        const accessToken = localStorage.getItem(Environment.LOCAL_STORAGE_KEY__ACCESS_TOKEN);
+        const token = `Bearer ${accessToken ? JSON.parse(accessToken) : ''}`;
+
+        const urlRelativa = `/pessoas?page=${page}&limit=${Environment.LIMITE_DE_LINHAS}&filter=${filter}`;
+        const { data, headers } = await Api.get(urlRelativa, { headers: { Authorization: token } });
 
         if(data){
             return {
@@ -41,7 +46,10 @@ const getAll = async (page = 1, filter = ''): Promise<TPessoasComTotalCount | Er
 
 const getById = async (id: number): Promise<IDetalhePessoa | Error> => {
     try {
-        const { data } = await Api.get(`/pessoas/${id}`);
+        const accessToken = localStorage.getItem(Environment.LOCAL_STORAGE_KEY__ACCESS_TOKEN);
+        const token = `Bearer ${accessToken ? JSON.parse(accessToken) : ''}`;
+
+        const { data } = await Api.get(`/pessoas/${id}`, { headers: { Authorization: token } });
 
         if(data){
             return data;
@@ -56,10 +64,13 @@ const getById = async (id: number): Promise<IDetalhePessoa | Error> => {
 
 const create = async (dados: Omit<IDetalhePessoa,'id'>): Promise<number | Error> => {
     try {
-        const { data } = await Api.post<IDetalhePessoa>('/pessoas', dados);
+        const accessToken = localStorage.getItem(Environment.LOCAL_STORAGE_KEY__ACCESS_TOKEN);
+        const token = `Bearer ${accessToken ? JSON.parse(accessToken) : ''}`;
+
+        const { data } = await Api.post('/pessoas', dados, { headers: { Authorization: token } });
 
         if(data){
-            return data.id;
+            return data;
         }
 
         return new Error('Erro ao criar registro.')
@@ -71,7 +82,10 @@ const create = async (dados: Omit<IDetalhePessoa,'id'>): Promise<number | Error>
 
 const updateById = async (id: number, dados: Omit<IDetalhePessoa, 'id'>): Promise<void | Error> => {
     try {
-        await Api.put(`/pessoas/${id}`, dados);
+        const accessToken = localStorage.getItem(Environment.LOCAL_STORAGE_KEY__ACCESS_TOKEN);
+        const token = `Bearer ${accessToken ? JSON.parse(accessToken) : ''}`;
+
+        await Api.put(`/pessoas/${id}`, dados, { headers: { Authorization: token } });
     } catch (error) {
         console.error(error);
         return new Error((error as {message: string}).message || 'Erro ao atualizar registro.');
@@ -80,7 +94,10 @@ const updateById = async (id: number, dados: Omit<IDetalhePessoa, 'id'>): Promis
 
 const deleteById = async (id: number): Promise<void | Error> => {
     try {
-        await Api.delete(`/pessoas/${id}`);
+        const accessToken = localStorage.getItem(Environment.LOCAL_STORAGE_KEY__ACCESS_TOKEN);
+        const token = `Bearer ${accessToken ? JSON.parse(accessToken) : ''}`;
+
+        await Api.delete(`/pessoas/${id}`, { headers: { Authorization: token } });
     } catch (error) {
         console.error(error);
         return new Error((error as {message: string}).message || 'Erro ao deletar registro.');
